style: Move Force to own header
