+++ include/CMakeLists.txt
@@ -25,6 +25,7 @@ set(DDG_HEADERS
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/mesh_io.h"
 
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/system.h"
+    "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/forces.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/mesh_mutator.h"
     "${CMAKE_CURRENT_SOURCE_DIR}/mem3dg/solver/trajfile.h"
