+++ include/mem3dg/solver/system.h
@@ -43,6 +43,7 @@
 #include "mem3dg/macros.h"
 #include "mem3dg/mesh_io.h"
 #include "mem3dg/meshops.h"
+#include "mem3dg/solver/forces.h"
 #include "mem3dg/solver/mesh_mutator.h"
 #include "mem3dg/type_utilities.h"
 
@@ -52,297 +53,6 @@ namespace gcs = ::geometrycentral::surface;
 namespace mem3dg {
 
 namespace solver {
-
-struct Forces {
-  /// Cached mesh of interest
-  gcs::ManifoldSurfaceMesh &mesh;
-  /// Embedding and other geometric details
-  gcs::VertexPositionGeometry &vpg;
-
-  /// Cached bending force
-  gcs::VertexData<double> bendingForce;
-  /// Cached tension-induced capillary force
-  gcs::VertexData<double> capillaryForce;
-  /// Cached interfacial line tension force
-  gcs::VertexData<double> lineCapillaryForce;
-  /// Cached externally-applied force
-  gcs::VertexData<double> externalForce;
-  /// Cached osmotic force
-  gcs::VertexData<double> osmoticForce;
-  /// Cached mechanical force
-  gcs::VertexData<double> mechanicalForce;
-  /// Cached osmotic pressure
-  double osmoticPressure;
-  /// Cached Surface tension
-  double surfaceTension;
-
-  /// Cached bending force
-  gcs::VertexData<gc::Vector3> bendingForceVec;
-  /// Cached bending force areaGrad component
-  gcs::VertexData<gc::Vector3> bendingForceVec_areaGrad;
-  /// Cached bending force gaussVec component
-  gcs::VertexData<gc::Vector3> bendingForceVec_gaussVec;
-  /// Cached bending force schlafliVec component
-  gcs::VertexData<gc::Vector3> bendingForceVec_schlafliVec;
-
-  /// Cached tension-induced capillary force
-  gcs::VertexData<gc::Vector3> capillaryForceVec;
-  /// Cached osmotic force
-  gcs::VertexData<gc::Vector3> osmoticForceVec;
-  /// Cached Dirichlet energy driven force
-  gcs::VertexData<gc::Vector3> lineCapillaryForceVec;
-  /// Cached adsorption driven force
-  gcs::VertexData<gc::Vector3> adsorptionForceVec;
-  /// Cached mechanical force
-  gcs::VertexData<gc::Vector3> mechanicalForceVec;
-
-  /// Cached local stretching forces (in-plane regularization)
-  gcs::VertexData<gc::Vector3> regularizationForce;
-  /// Cached damping forces
-  gcs::VertexData<gc::Vector3> dampingForce;
-  /// Cached stochastic forces
-  gcs::VertexData<gc::Vector3> stochasticForce;
-
-  /// Cached interior penalty chemical potential
-  gcs::VertexData<double> interiorPenaltyPotential;
-  /// Cached bending related chemical potential
-  gcs::VertexData<double> bendingPotential;
-  /// Cached adsorption related chemical potential
-  gcs::VertexData<double> adsorptionPotential;
-  /// Cached dirichlet energy related chemical potential
-  gcs::VertexData<double> diffusionPotential;
-  /// Cached chemical potential
-  gcs::VertexData<double> chemicalPotential;
-
-  /// force mask
-  gcs::VertexData<gc::Vector3> forceMask;
-  /// protein mask
-  gcs::VertexData<double> proteinMask;
-
-  Forces(gcs::ManifoldSurfaceMesh &mesh_, gcs::VertexPositionGeometry &vpg_)
-      : mesh(mesh_), vpg(vpg_), mechanicalForce(mesh, 0),
-        mechanicalForceVec(mesh, {0, 0, 0}), bendingForceVec(mesh, {0, 0, 0}),
-        bendingForceVec_areaGrad(mesh, {0, 0, 0}),
-        bendingForceVec_gaussVec(mesh, {0, 0, 0}),
-        bendingForceVec_schlafliVec(mesh, {0, 0, 0}),
-        capillaryForceVec(mesh, {0, 0, 0}), osmoticForceVec(mesh, {0, 0, 0}),
-        adsorptionForceVec(mesh, {0, 0, 0}),
-        lineCapillaryForceVec(mesh, {0, 0, 0}), bendingForce(mesh, 0),
-        capillaryForce(mesh, 0), surfaceTension(0), lineCapillaryForce(mesh, 0),
-        externalForce(mesh, 0), osmoticForce(mesh, 0), osmoticPressure(0),
-        regularizationForce(mesh, {0, 0, 0}), stochasticForce(mesh, {0, 0, 0}),
-        dampingForce(mesh, {0, 0, 0}), interiorPenaltyPotential(mesh, 0),
-        bendingPotential(mesh, 0), adsorptionPotential(mesh, 0),
-        diffusionPotential(mesh, 0), chemicalPotential(mesh, 0),
-        forceMask(mesh, {1.0, 1.0, 1.0}), proteinMask(mesh, 1) {}
-
-  ~Forces() {}
-
-  // ==========================================================
-  // =============      Data interop helpers    ===============
-  // ==========================================================
-
-  // /**
-  //  * @brief Return constructed vertexData that from Eigen matrix or vector
-  //  */
-  // gcs::VertexData<double> toVertexData(EigenVectorX1d &vector) {
-  //   return gcs::VertexData<double>(mesh, vector);
-  // }
-
-  // gcs::VertexData<double> toVertexData(EigenVectorX1d &&vector) {
-  //   return gcs::VertexData<double>(mesh, vector);
-  // }
-
-  // gcs::VertexData<gc::Vector3> toVertexData(EigenVectorX3dr &vector) {
-  //   gcs::VertexData<gc::Vector3> vertexData(mesh);
-  //   gc::EigenMap<double, 3>(vertexData) = vector;
-  //   return vertexData;
-  // }
-
-  // gcs::VertexData<gc::Vector3> toVertexData(EigenVectorX3dr &&vector) {
-  //   gcs::VertexData<gc::Vector3> vertexData(mesh);
-  //   gc::EigenMap<double, 3>(vertexData) = vector;
-  //   return vertexData;
-  // }
-
-  /**
-   * @brief Find the corresponding scalar vector (vertexData) by rowwise
-   * (vertexwise) projecting matrix (vector vertexData) onto angle-weighted
-   * normal vector
-   */
-  gcs::VertexData<double>
-  ontoNormal(const gcs::VertexData<gc::Vector3> &vector) const {
-    gcs::VertexData<double> vertexData(
-        mesh, rowwiseDotProduct(gc::EigenMap<double, 3>(vector),
-                                gc::EigenMap<double, 3>(vpg.vertexNormals)));
-    return vertexData;
-  }
-  gcs::VertexData<double>
-  ontoNormal(const gcs::VertexData<gc::Vector3> &&vector) const {
-    gcs::VertexData<double> vertexData(
-        mesh, rowwiseDotProduct(gc::EigenMap<double, 3>(vector),
-                                gc::EigenMap<double, 3>(vpg.vertexNormals)));
-    return vertexData;
-  }
-
-  EigenVectorX1d ontoNormal(const EigenVectorX3dr &vector) const {
-    return rowwiseDotProduct(vector,
-                             gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-  EigenVectorX1d ontoNormal(const EigenVectorX3dr &&vector) const {
-    return rowwiseDotProduct(vector,
-                             gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-  double ontoNormal(const gc::Vector3 &vector, const gc::Vertex &v) const {
-    return gc::dot(vector, vpg.vertexNormals[v]);
-  }
-
-  double ontoNormal(const gc::Vector3 &vector, const std::size_t i) const {
-    return gc::dot(vector, vpg.vertexNormals[i]);
-  }
-
-  double ontoNormal(const gc::Vector3 &&vector, gc::Vertex &v) const {
-    return gc::dot(vector, vpg.vertexNormals[v]);
-  }
-
-  /**
-   * @brief Find the corresponding matrix (vector vertexData) by rowwise
-   * (vertexwise) appending angle-weighted normal vector to the scalar vector
-   * (vertexData)
-   */
-  gcs::VertexData<gc::Vector3> addNormal(gcs::VertexData<double> &vector) {
-    gcs::VertexData<gc::Vector3> vertexData(mesh);
-    gc::EigenMap<double, 3>(vertexData) = rowwiseScalarProduct(
-        vector.raw(), gc::EigenMap<double, 3>(vpg.vertexNormals));
-    return vertexData;
-  }
-  gcs::VertexData<gc::Vector3> addNormal(gcs::VertexData<double> &&vector) {
-    gcs::VertexData<gc::Vector3> vertexData(mesh);
-    gc::EigenMap<double, 3>(vertexData) = rowwiseScalarProduct(
-        vector.raw(), gc::EigenMap<double, 3>(vpg.vertexNormals));
-    return vertexData;
-  }
-
-  EigenVectorX3dr addNormal(EigenVectorX1d &vector) {
-    return rowwiseScalarProduct(vector,
-                                gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-  EigenVectorX3dr addNormal(EigenVectorX1d &&vector) {
-    return rowwiseScalarProduct(vector,
-                                gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-
-  gc::Vector3 addNormal(double &vector, gc::Vertex &v) {
-    return vector * vpg.vertexNormals[v];
-  }
-  gc::Vector3 addNormal(double &&vector, gc::Vertex &v) {
-    return vector * vpg.vertexNormals[v];
-  }
-
-  /**
-   * @brief Project the vector onto tangent plane by removing the
-   * angle-weighted normal component
-   */
-  EigenVectorX3dr toTangent(gcs::VertexData<gc::Vector3> &vector) {
-    return gc::EigenMap<double, 3>(vector) -
-           rowwiseScalarProduct(
-               rowwiseDotProduct(gc::EigenMap<double, 3>(vector),
-                                 gc::EigenMap<double, 3>(vpg.vertexNormals)),
-               gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-  EigenVectorX3dr toTangent(gcs::VertexData<gc::Vector3> &&vector) {
-    return gc::EigenMap<double, 3>(vector) -
-           rowwiseScalarProduct(
-               rowwiseDotProduct(gc::EigenMap<double, 3>(vector),
-                                 gc::EigenMap<double, 3>(vpg.vertexNormals)),
-               gc::EigenMap<double, 3>(vpg.vertexNormals));
-  }
-
-  gc::Vector3 toTangent(gc::Vector3 &vector, gc::Vertex &v) {
-    return vector -
-           gc::dot(vector, vpg.vertexNormals[v]) * vpg.vertexNormals[v];
-  }
-  gc::Vector3 toTangent(gc::Vector3 &&vector, gc::Vertex &v) {
-    return vector -
-           gc::dot(vector, vpg.vertexNormals[v]) * vpg.vertexNormals[v];
-  }
-
-  /**
-   * @brief Find the masked force
-   */
-  gcs::VertexData<gc::Vector3> maskForce(gcs::VertexData<gc::Vector3> &vector) {
-    gcs::VertexData<gc::Vector3> vertexData(mesh);
-    gc::EigenMap<double, 3>(vertexData).array() =
-        gc::EigenMap<double, 3>(vector).array() *
-        gc::EigenMap<double, 3>(forceMask).array();
-    return vertexData;
-  }
-
-  gcs::VertexData<gc::Vector3>
-  maskForce(gcs::VertexData<gc::Vector3> &&vector) {
-    gcs::VertexData<gc::Vector3> vertexData(mesh);
-    gc::EigenMap<double, 3>(vertexData).array() =
-        gc::EigenMap<double, 3>(vector).array() *
-        gc::EigenMap<double, 3>(forceMask).array();
-    return vertexData;
-  }
-
-  EigenVectorX3dr maskForce(const EigenVectorX3dr &vector) const {
-    return vector.array() * gc::EigenMap<double, 3>(forceMask).array();
-  }
-
-  EigenVectorX3dr maskForce(const EigenVectorX3dr &&vector) const {
-    return vector.array() * gc::EigenMap<double, 3>(forceMask).array();
-  }
-
-  gc::Vector3 maskForce(const gc::Vector3 &vector, const gc::Vertex &v) const {
-    return gc::Vector3{vector.x * forceMask[v].x, vector.y * forceMask[v].y,
-                       vector.z * forceMask[v].z};
-  }
-
-  gc::Vector3 maskForce(const gc::Vector3 &vector, const std::size_t i) const {
-    return gc::Vector3{vector.x * forceMask[i].x, vector.y * forceMask[i].y,
-                       vector.z * forceMask[i].z};
-  }
-
-  gc::Vector3 maskForce(const gc::Vector3 &&vector, const gc::Vertex &v) const {
-    return gc::Vector3{vector.x * forceMask[v].x, vector.y * forceMask[v].y,
-                       vector.z * forceMask[v].z};
-  }
-
-  /**
-   * @brief Find the masked chemical potential
-   */
-  gcs::VertexData<double> maskProtein(gcs::VertexData<double> &potential) {
-    gcs::VertexData<double> vertexData(mesh);
-    toMatrix(vertexData).array() =
-        toMatrix(potential).array() * toMatrix(proteinMask).array();
-    return vertexData;
-  }
-
-  gcs::VertexData<double> maskProtein(gcs::VertexData<double> &&potential) {
-    gcs::VertexData<double> vertexData(mesh);
-    toMatrix(vertexData).array() =
-        toMatrix(potential).array() * toMatrix(proteinMask).array();
-    return vertexData;
-  }
-
-  EigenVectorX1d maskProtein(EigenVectorX1d &potential) {
-    return potential.array() * toMatrix(proteinMask).array();
-  }
-
-  EigenVectorX1d maskProtein(EigenVectorX1d &&potential) {
-    return potential.array() * toMatrix(proteinMask).array();
-  }
-
-  double maskProtein(double &potential, gc::Vertex &v) {
-    return potential * proteinMask[v];
-  }
-
-  double maskProtein(double &&potential, gc::Vertex &v) {
-    return potential * proteinMask[v];
-  }
-};
 
 struct Parameters {
   /// Bending modulus
@@ -467,8 +177,6 @@ protected:
   gcs::FaceData<double> refFaceAreas;
   /// Cached geodesic distance
   gcs::VertexData<double> geodesicDistanceFromPtInd;
-  // /// V-E distribution matrix
-  // Eigen::SparseMatrix<double> D;
   /// Random number engine
   pcg32 rng;
   std::normal_distribution<double> normal_dist;
